+++ modules/claims_api/lib/bgs_service/vnp_atchms_service.rb
@@ -2,6 +2,8 @@
 
 module ClaimsApi
   class VnpAtchmsService < ClaimsApi::LocalBGS
+    include ClaimsApi::BGSHelpers
+
     # Takes an object with a minimum of (other fields are camelized and passed to BGS):
     # vnp_proc_id: BGS procID
     # atchms_file_nm: File name
