+++ modules/claims_api/lib/bgs_service/vnp_person_service.rb
@@ -2,6 +2,8 @@
 
 module ClaimsApi
   class VnpPersonService < ClaimsApi::LocalBGS
+    include ClaimsApi::BGSHelpers
+
     # Takes an object with a minimum of (other fields are camelized and passed to BGS):
     # vnp_proc_id: BGS procID
     # vnp_ptcpnt_id: Veteran's participant id
