+++ modules/claims_api/spec/lib/claims_api/local_bgs_spec.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
 require 'rails_helper'
-require 'bgs_service/local_bgs'
+require 'bgs_service/miscellaneous_bgs_operations'
 require 'claims_api/error/soap_error_handler'
 
-describe ClaimsApi::LocalBGS do
+describe ClaimsApi::MiscellaneousBGSOperations do
   subject { described_class.new external_uid: 'xUid', external_key: 'xKey' }
 
   let(:soap_error_handler) { ClaimsApi::SoapErrorHandler.new }
@@ -12,11 +12,9 @@ describe ClaimsApi::LocalBGS do
   describe '#find_poa_by_participant_id' do
     it 'responds as expected, with extra ClaimsApi::Logger logging' do
       VCR.use_cassette('claims_api/bgs/claimant_web_service/find_poa_by_participant_id') do
-        allow_any_instance_of(BGS::OrgWebService).to receive(:find_poa_history_by_ptcpnt_id).and_return({})
-
         # Events logged:
-        # 1: establish_ssl_connection - how long to establish the connection
-        # 2: connection_wsdl_get - duration of WSDL request cycle
+        # 1: connection_wsdl_get - duration of WSDL request cycle
+        # 2: built_request - how long to build the request
         # 3: connection_post - how long does the post itself take for the request cycle
         # 4: parsed_response - how long to parse the response
         expect(ClaimsApi::Logger).to receive(:log).exactly(4).times
@@ -44,9 +42,7 @@ describe ClaimsApi::LocalBGS do
     it 'triggers StatsD measurements' do
       VCR.use_cassette('claims_api/bgs/claimant_web_service/find_poa_by_participant_id',
                        allow_playback_repeats: true) do
-        allow_any_instance_of(BGS::OrgWebService).to receive(:find_poa_history_by_ptcpnt_id).and_return({})
-
-        %w[establish_ssl_connection connection_wsdl_get connection_post parsed_response].each do |event|
+        %w[connection_wsdl_get built_request connection_post parsed_response].each do |event|
           expect { subject.find_poa_by_participant_id('does-not-matter') }
             .to trigger_statsd_measure("api.claims_api.local_bgs.#{event}.duration")
         end
@@ -107,9 +103,9 @@ describe ClaimsApi::LocalBGS do
                                                                ),
                                                                key: 'PersonDTO').and_return(:bgs_unknown_error_message)
         begin
-          allow(soap_error_handler).to receive(:handle_errors)
+          allow(soap_error_handler).to receive(:handle_errors!)
             .with(:bgs_unknown_error_message).and_raise(Common::Exceptions::UnprocessableEntity)
-          ret = soap_error_handler.send(:handle_errors, :bgs_unknown_error_message)
+          ret = soap_error_handler.send(:handle_errors!, :bgs_unknown_error_message)
           expect(ret.class).to_be Array
           expect(ret.size).to eq 1
         rescue => e
