+++ modules/claims_api/spec/requests/v1/intent_to_file_request_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 require_relative '../../rails_helper'
-require 'bgs_service/local_bgs'
+require 'bgs_service/miscellaneous_bgs_operations'
 
 RSpec.describe 'Intent to file', type: :request do
   let(:headers) do
@@ -151,7 +151,7 @@ RSpec.describe 'Intent to file', type: :request do
       describe 'handling the claimant fields' do
         context "when 'participant_claimant_id' is provided" do
           it 'that field and value are sent to BGS' do
-            expect_any_instance_of(ClaimsApi::LocalBGS)
+            expect_any_instance_of(ClaimsApi::MiscellaneousBGSOperations)
               .to receive(:insert_intent_to_file).with(hash_including(participant_claimant_id: '123')).and_return({})
 
             mock_acg(scopes) do |auth_header|
@@ -163,7 +163,7 @@ RSpec.describe 'Intent to file', type: :request do
 
         context "when 'claimant_ssn' is provided" do
           it 'that field and value are sent to BGS' do
-            expect_any_instance_of(ClaimsApi::LocalBGS)
+            expect_any_instance_of(ClaimsApi::MiscellaneousBGSOperations)
               .to receive(:insert_intent_to_file).with(hash_including(claimant_ssn: '123')).and_return({})
 
             mock_acg(scopes) do |auth_header|
@@ -179,7 +179,7 @@ RSpec.describe 'Intent to file', type: :request do
           end
 
           it "'participant_claimant_id' is set to the target_veteran.participant_id and sent to BGS " do
-            expect_any_instance_of(ClaimsApi::LocalBGS)
+            expect_any_instance_of(ClaimsApi::MiscellaneousBGSOperations)
               .to receive(:insert_intent_to_file).with(hash_including(participant_claimant_id: '999')).and_return({})
 
             mock_acg(scopes) do |auth_header|
@@ -190,7 +190,7 @@ RSpec.describe 'Intent to file', type: :request do
 
         context "when both 'participant_claimant_id' and 'claimant_ssn' are provided" do
           it "both 'participant_claimant_id' and 'claimant_ssn' are sent to BGS " do
-            expect_any_instance_of(ClaimsApi::LocalBGS)
+            expect_any_instance_of(ClaimsApi::MiscellaneousBGSOperations)
               .to receive(:insert_intent_to_file).with(
                 hash_including(
                   participant_claimant_id: '123', claimant_ssn: '456'
