+++ modules/claims_api/spec/requests/v1/rswag_claims_request_spec.rb
@@ -127,7 +127,7 @@ describe 'EVSS Claims management', openapi_spec: 'modules/claims_api/app/swagger
           before do |example|
             stub_poa_verification
 
-            allow_any_instance_of(ClaimsApi::LocalBGS).to receive(:all).and_raise(
+            allow_any_instance_of(ClaimsApi::MiscellaneousBGSOperations).to receive(:all).and_raise(
               Common::Exceptions::ResourceNotFound.new(detail: 'The Resource was not found.')
             )
             mock_acg(scopes) do
