+++ modules/claims_api/spec/requests/v1/rswag_power_of_attorney_request_spec.rb
@@ -7,7 +7,7 @@ require_relative '../../rails_helper'
 require_relative '../../support/swagger_shared_components/v1'
 
 describe 'Power of Attorney', openapi_spec: 'modules/claims_api/app/swagger/claims_api/v1/swagger.json' do # rubocop:disable RSpec/DescribeClass
-  let(:pws) { ClaimsApi::LocalBGS }
+  let(:pws) { ClaimsApi::MiscellaneousBGSOperations }
 
   path '/forms/2122' do
     get 'Gets schema for POA form.' do
