+++ modules/veteran/spec/models/veteran/user_spec.rb
@@ -6,7 +6,7 @@ describe Veteran::User do
   context 'initialization' do
     let(:user) { FactoryBot.create(:user, :loa3) }
 
-    let(:ows) { ClaimsApi::LocalBGS }
+    let(:ows) { ClaimsApi::MiscellaneousBGSOperations }
 
     it 'initializes from a user' do
       VCR.use_cassette('bgs/claimant_web_service/find_poa_by_participant_id') do
